requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0
apify-client>=1.6.0
